finesse the csv datatypes better.  plus rehome the show() effects
