+++ seventyfive/pom.xml
@@ -8,12 +8,7 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-<description>75 pandas excercises
-
-
-
-</description>
-
+    <description>75 pandas excercises</description>
     <dependencies>
         <dependency>
             <groupId>columnar</groupId>
@@ -22,9 +17,5 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
-
-
     <artifactId>seventyfive</artifactId>
-
-
 </project>
