+++ vector-like/pom.xml
@@ -2,25 +2,11 @@
 <project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <source>14</source>
-                    <target>14</target>
-                    <compilerArgs>--enable-preview</compilerArgs>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
     <parent>
         <artifactId>columnar-parent</artifactId>
         <groupId>columnar</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
-
     <dependencies>
         <dependency>
             <groupId>org.bouncycastle</groupId>
@@ -28,6 +14,5 @@
             <version>1.64</version>
         </dependency>
     </dependencies>
-
- <artifactId>vector-like</artifactId>
+    <artifactId>vector-like</artifactId>
 </project>
